move is_eigen_row_vector to scale so it knows it exists
